refactored sorting
